fancy build logic

+++ Rakefile
@@ -1,3 +1,4 @@
+require 'fileutils'
 require 'rspec/core/rake_task'
 require 'rubocop/rake_task'
 
@@ -6,38 +7,43 @@ RSpec::Core::RakeTask.new(:spec) { |t|
   t.pattern = Dir.glob('spec/**/*_spec.rb')
 }
 
-namespace(:gem) {
-  desc 'Bump gem minor version'
-  task(:bump_version) {
-    Rake::FileList.new('*.gemspec').each { |file|
-      lines = File.readlines(file)
-      lines.map! { |line|
-        if line.strip.start_with?('spec.version')
-          version = line.strip.split('=').last.strip.gsub(/['|"]/, '')
-          main_version, sub_version = version.split('.').map(&:to_i)
-          line.gsub(version, "#{main_version}.#{sub_version + 1}")
-        else
-          line
-        end
-      }
-      File.write(file, lines.join)
-    }
-  }
+def bump_version
+  file = Rake::FileList.new('*.gemspec').first
+  mtime = [File.mtime('lib'), File.mtime('sig'), File.mtime(file)].max
+  if File.exist?('.gt') && File.mtime('.gt') >= mtime
+    puts 'No files have changed since latest gem install'
+    return false
+  end
 
-  desc 'Bump gem minor version and build'
-  task(build: :bump_version) {
-    Rake::FileList.new('*.gemspec').each { |file|
-      sh "gem build #{file}"
-    }
+  lines = File.readlines(file)
+  lines.map! { |line|
+    if line.strip.start_with?('spec.version')
+      version = line.strip.split('=').last.strip.gsub(/['|"]/, '')
+      main_version, sub_version = version.split('.').map(&:to_i)
+      line.gsub(version, "#{main_version}.#{sub_version + 1}")
+    else
+      line
+    end
   }
+  File.write(file, lines.join)
+  FileUtils.touch('.gt')
+end
 
-  desc 'Bump gem minor version, build, and install'
-  task(install: :build) {
-    Rake::FileList.new('*.gem').each { |file|
-      sh "gem install #{file}"
-      sh "rm #{file}"
-    }
-  }
+def build_gem
+  return false unless bump_version
+
+  file = Rake::FileList.new('*.gemspec').first
+  sh "gem build #{file}"
+  return true
+end
+
+desc 'Bump gem minor version, build, and install'
+task(:install) {
+  next unless build_gem
+
+  file = Rake::FileList.new('*.gem').first
+  sh "gem install #{file}"
+  sh "rm #{file}"
 }
 
 task default: %w[rubocop spec]
